+++ server/src/test/java/org/elasticsearch/search/query/ThrowingQueryBuilder.java
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.xpack.search;
+package org.elasticsearch.search.query;
 
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
@@ -21,7 +21,8 @@ import org.elasticsearch.xcontent.XContentBuilder;
 
 import java.io.IOException;
 
-class ThrowingQueryBuilder extends AbstractQueryBuilder<ThrowingQueryBuilder> {
+// moved from x-pack to server module
+public class ThrowingQueryBuilder extends AbstractQueryBuilder<ThrowingQueryBuilder> {
     public static final String NAME = "throw";
 
     private final long randomUID;
@@ -36,7 +37,7 @@ class ThrowingQueryBuilder extends AbstractQueryBuilder<ThrowingQueryBuilder> {
      * @param failure what exception to throw
      * @param shardId what shardId to throw the exception. If shardId is less than 0, it will throw for all shards.
      */
-    ThrowingQueryBuilder(long randomUID, RuntimeException failure, int shardId) {
+    public ThrowingQueryBuilder(long randomUID, RuntimeException failure, int shardId) {
         super();
         this.randomUID = randomUID;
         this.failure = failure;
@@ -51,7 +52,7 @@ class ThrowingQueryBuilder extends AbstractQueryBuilder<ThrowingQueryBuilder> {
      * @param failure what exception to throw
      * @param index what index to throw the exception against (all shards of that index)
      */
-    ThrowingQueryBuilder(long randomUID, RuntimeException failure, String index) {
+    public ThrowingQueryBuilder(long randomUID, RuntimeException failure, String index) {
         super();
         this.randomUID = randomUID;
         this.failure = failure;
@@ -59,7 +60,7 @@ class ThrowingQueryBuilder extends AbstractQueryBuilder<ThrowingQueryBuilder> {
         this.index = index;
     }
 
-    ThrowingQueryBuilder(StreamInput in) throws IOException {
+    public ThrowingQueryBuilder(StreamInput in) throws IOException {
         super(in);
         this.randomUID = in.readLong();
         this.failure = in.readException();
